Fix regex check for valid RSA key
Generating a key pair at https://www.transip.nl/cp/mijn-account/#api
yields a private key with a `BEGIN RSA PRIVATE KEY` header, as opposed to
the `BEGIN PRIVATE KEY` header the regex checks for.

Since there are probably still valid keys in use that include the `RSA`
part in the header, I made it optional (as opposed to removing it
entirely).

It might be a good idea to consider not doing the check in the first place.